move salary-info changeStatus to root
